Interfaz Grafica3
Se implemento de forma casi total la interfaz de secretaria(falta las tablas y creación relacionadas con edificios). Incluyendo ponerle el menú de los vehículos a la barra de menú y toda la implementación correspondiente.

Se cambió el atributo del Carnet de identidad a long.

Se creó el método getTrabajadoresPorTipo en EmpresaAcopio, además se refactorizó la clase entera, borrando métodos innecesarios  y organizando su orden.

Se añadieron getters y setters al id de los vehículos, también se le dio su método toString().

+++ Aprovisionamiento De Leche/src/Personas/Persona.java
@@ -3,11 +3,11 @@ package Personas;
 public class Persona {
 
     private String Nombre;
-    private int CI;
+    private long CI;
     private String Apellidos;
     
     public Persona(){}
-    public Persona(String nombre, String apellidos, int CI){
+    public Persona(String nombre, String apellidos, long CI){
         this.Nombre=nombre;
         this.Apellidos=apellidos;
         this.CI=CI;
@@ -25,7 +25,7 @@ public class Persona {
         this.Nombre=Nombre;
     }
 
-    public int getCI() {
+    public long getCI() {
         return this.CI;
     }
 
@@ -33,7 +33,7 @@ public class Persona {
      * 
      * @param CI
      */
-    public void setCI(int CI) {
+    public void setCI(long CI) {
         this.CI=CI;
     }
